+++ src/main/kotlin/app/revanced/patches/youtube/misc/mapping/patch/ResourceIdMappingProviderResourcePatch.kt
@@ -4,10 +4,10 @@ import app.revanced.extensions.doRecursively
 import app.revanced.patcher.annotation.Description
 import app.revanced.patcher.annotation.Name
 import app.revanced.patcher.annotation.Version
-import app.revanced.patcher.data.implementation.ResourceData
-import app.revanced.patcher.patch.implementation.ResourcePatch
-import app.revanced.patcher.patch.implementation.misc.PatchResult
-import app.revanced.patcher.patch.implementation.misc.PatchResultSuccess
+import app.revanced.patcher.data.impl.ResourceData
+import app.revanced.patcher.patch.PatchResult
+import app.revanced.patcher.patch.PatchResultSuccess
+import app.revanced.patcher.patch.impl.ResourcePatch
 import org.w3c.dom.Element
 
 @Name("resource-id-mapping-provider-resource-patch-dependency")
@@ -20,7 +20,7 @@ class ResourceIdMappingProviderResourcePatch : ResourcePatch() {
     }
 
     override fun execute(data: ResourceData): PatchResult {
-        data.getXmlEditor("res/values/public.xml").use { editor ->
+        data.xmlEditor["res/values/public.xml"].use { editor ->
             resourceMappings = buildMap {
                 editor.file.documentElement.doRecursively { node ->
                     if (node !is Element) return@doRecursively