+++ src/main/kotlin/app/revanced/patches/youtube/misc/microg/fingerprints/IntegrityCheckFingerprint.kt
@@ -1,0 +1,82 @@
+package app.revanced.patches.youtube.misc.microg.fingerprints
+
+import app.revanced.patcher.annotation.Name
+import app.revanced.patcher.annotation.Version
+import app.revanced.patcher.extensions.or
+import app.revanced.patcher.fingerprint.method.impl.MethodFingerprint
+import app.revanced.patcher.fingerprint.method.annotation.FuzzyPatternScanMethod
+import app.revanced.patcher.fingerprint.method.annotation.MatchingMethod
+import app.revanced.patches.youtube.misc.microg.annotations.MicroGPatchCompatibility
+import org.jf.dexlib2.AccessFlags
+import org.jf.dexlib2.Opcode
+
+@Name("google-play-sig-check-fingerprint")
+@MatchingMethod(
+    "Ldwn;", "d"
+)
+@FuzzyPatternScanMethod(2) // FIXME: Test this threshold and find the best value.
+@MicroGPatchCompatibility
+@Version("0.0.1")
+object IntegrityCheckFingerprint : MethodFingerprint(
+    "L", AccessFlags.PUBLIC or AccessFlags.STATIC, listOf("L", "L"), listOf(
+        Opcode.MOVE_OBJECT_FROM16,
+        Opcode.CONST_STRING,
+        Opcode.CONST_STRING,
+        Opcode.NEW_INSTANCE,
+        Opcode.MOVE_OBJECT_FROM16,
+        Opcode.INVOKE_DIRECT,
+        Opcode.CONST_4,
+        Opcode.CONST_STRING,
+        Opcode.INVOKE_STATIC,
+        Opcode.MOVE_RESULT_OBJECT,
+        Opcode.INVOKE_STATIC,
+        Opcode.MOVE_RESULT_OBJECT,
+        Opcode.IF_EQ,
+        Opcode.IGET_OBJECT,
+        Opcode.INVOKE_VIRTUAL,
+        Opcode.SGET,
+        Opcode.IGET_OBJECT,
+        Opcode.INVOKE_STATIC,
+        Opcode.MOVE_RESULT_WIDE,
+        Opcode.CONST_WIDE,
+        Opcode.CONST_STRING,
+        Opcode.CONST_4,
+        Opcode.CONST_STRING,
+        Opcode.CONST_4,
+        Opcode.CMP_LONG,
+        Opcode.IF_GEZ,
+        Opcode.CONST_16,
+        Opcode.GOTO_16,
+        Opcode.CONST_STRING,
+        Opcode.IGET_OBJECT,
+        Opcode.INVOKE_VIRTUAL,
+        Opcode.MOVE_RESULT_OBJECT,
+        Opcode.CONST_16,
+        Opcode.INVOKE_VIRTUAL,
+        Opcode.MOVE_RESULT_OBJECT,
+        Opcode.GOTO,
+        Opcode.MOVE_EXCEPTION,
+        Opcode.CONST_STRING,
+        Opcode.INVOKE_STATIC,
+        Opcode.CONST_4,
+        Opcode.IGET_OBJECT,
+        Opcode.IF_EQZ,
+        Opcode.CONST_STRING,
+        Opcode.INVOKE_VIRTUAL,
+        Opcode.MOVE_RESULT,
+        Opcode.CONST,
+        Opcode.IF_NE,
+        Opcode.CONST_16,
+        Opcode.INVOKE_VIRTUAL,
+        Opcode.MOVE_RESULT_OBJECT,
+        Opcode.INVOKE_VIRTUAL,
+        Opcode.MOVE_RESULT_OBJECT,
+        Opcode.SGET_OBJECT,
+        Opcode.INVOKE_STATIC,
+        Opcode.MOVE_RESULT_OBJECT,
+        Opcode.IF_NEZ,
+        Opcode.CONST_STRING,
+        Opcode.GOTO,
+        Opcode.NEW_ARRAY
+    ), listOf("This should never happen.", "GooglePlayServicesUtil", "Google Play Store signature invalid.")
+)